Filter linter results based on the current file
Some linters can not process a single file in isolation so they only
work for directories processing all files and returning their respective
linting errors. All linting errors were then displayed with a diagnostic
sign in the signcolumn of the current file even though they may not
belong to that file they are displayed in.